refactor(skills): extract uninstallSkillLocally to its own module, cutting the CLI artery
catalog-install is reachable from skills/load (skill_load tool) and imported
deleteSkillCapabilityNode from capability-seed, which statically imports
buildCliProgram (cli/program). Move uninstallSkillLocally — the only caller — to
src/skills/uninstall.ts so catalog-install no longer imports capability-seed.
deleteSkillCapabilityNode stays in capability-seed.

Tool-manifest eager import graph: 705 -> 318 modules; the skills/load ->
catalog-install -> capability-seed -> cli/program artery is gone.
